+++ config/routes.rb
@@ -42,6 +42,7 @@ Rails.application.routes.draw do
     member do
       get "spending_breakdown_download"
     end
+    get "continuing_activities", on: :collection
   end
 
   namespace :exports do