+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   comments
 - the new actual, refund and activity importer no longer accepts negative actual
   values
+- Provide a CSV download of all activities that are likely to continue under DSIT and will need the new DSIT transparency identifier
 
 ## Release 141 - 2023-12-04